moved files below 'test/spec/adapt' files to 'test/spec/vivliostyle' directory
